refactor(markdown): update components category for MarkdownRenderer (#150)
